pyproject.toml
[tool.poetry]
name = "loan-backend"
version = "0.1.0"
description = "Django-based loan application backend"
authors = ["Tare Appah <grillzwitu@gmail.com>"]
license = "None"
readme = "README.md"
packages = [
    { include = "loan_app" },
    { include = "users" },
    { include = "loan" },
    { include = "fraud" }
]

[tool.poetry.dependencies]
python = ">=3.12,<4"
Django = ">=4.2"
djangorestframework = ">=3.14"
django-environ = ">=0.9.0"
django-cors-headers = ">=4.0,<5"
django-redis = ">=5.2"
psycopg2-binary = ">=2.9"
drf-yasg = ">=1.21.5"
djangorestframework-simplejwt = ">=5.2.2"
whitenoise = ">=6.0"
gunicorn = ">=22.0.0,<24.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-django = ">=4.5"
pytest-cov = ">=4.1"
flake8 = ">=5.0"
mypy = ">=1.2"
black = "^25.1.0"
isort = "^6.0.1"
django-stubs = "^5.2.1"
djangorestframework-stubs = "^3.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "loan_app.settings"
strict_settings = true

[tool.flake8]
max-line-length = 88
exclude = ["*/migrations/*"]
